GpgKey should be internal
